Reorganize packages and structs
