+++ internal/service-manager/client_test.go
@@ -3,36 +3,21 @@ package servicemanager_test
 import (
 	"context"
 	"encoding/json"
-	"fmt"
-	"log"
 	"log/slog"
-	"net/http"
-	"net/http/httptest"
-	"os"
-	"strings"
 	"testing"
 
-	"github.com/google/uuid"
+	clusterobject "github.com/kyma-project/btp-manager/internal/cluster-object"
 	servicemanager "github.com/kyma-project/btp-manager/internal/service-manager"
 	"github.com/kyma-project/btp-manager/internal/service-manager/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-const (
-	serviceOfferingsJSONPath = "testdata/service_offerings.json"
-	servicePlansJSONPath     = "testdata/service_plans.json"
-	serviceInstancesJSONPath = "testdata/service_instances.json"
-	serviceBindingsJSONPath  = "testdata/service_bindings.json"
 )
 
 func TestClient(t *testing.T) {
 	// given
-	secretProvider := newFakeSecretProvider()
-	secretProvider.AddSecret(defaultSecret())
-	srv, err := initFakeServer()
+	secretProvider := clusterobject.NewFakeSecretProvider()
+	secretProvider.AddSecret(clusterobject.FakeDefaultSecret())
+	srv, err := servicemanager.NewFakeServer()
 	require.NoError(t, err)
 
 	srv.Start()
@@ -40,13 +25,13 @@ func TestClient(t *testing.T) {
 	httpClient := srv.Client()
 	url := srv.URL
 
-	defaultServiceOfferings, err := getServiceOfferingsFromJSON()
+	defaultServiceOfferings, err := servicemanager.GetServiceOfferingsFromJSON()
 	require.NoError(t, err)
-	defaultServicePlans, err := getServicePlansFromJSON()
+	defaultServicePlans, err := servicemanager.GetServicePlansFromJSON()
 	require.NoError(t, err)
-	defaultServiceInstances, err := getServiceInstancesFromJSON()
+	defaultServiceInstances, err := servicemanager.GetServiceInstancesFromJSON()
 	require.NoError(t, err)
-	defaultServiceBindings, err := getServiceBindingsFromJSON()
+	defaultServiceBindings, err := servicemanager.GetServiceBindingsFromJSON()
 	require.NoError(t, err)
 
 	t.Run("should get service offerings available for the default credentials", func(t *testing.T) {
@@ -190,7 +175,7 @@ func TestClient(t *testing.T) {
 		assert.Equal(t, expectedParams, actualParams)
 
 		// revert server to default state
-		srv, err = initFakeServer()
+		srv, err = servicemanager.NewFakeServer()
 		require.NoError(t, err)
 	})
 
@@ -216,7 +201,7 @@ func TestClient(t *testing.T) {
 		assert.Equal(t, siUpdatedShared, si.Shared)
 
 		// revert server to default state
-		srv, err = initFakeServer()
+		srv, err = servicemanager.NewFakeServer()
 		require.NoError(t, err)
 	})
 
@@ -241,7 +226,7 @@ func TestClient(t *testing.T) {
 		require.Error(t, err)
 
 		// revert server to default state
-		srv, err = initFakeServer()
+		srv, err = servicemanager.NewFakeServer()
 		require.NoError(t, err)
 	})
 
@@ -335,547 +320,9 @@ func TestClient(t *testing.T) {
 		require.Error(t, err)
 
 		// revert server to default state
-		srv, err = initFakeServer()
+		srv, err = servicemanager.NewFakeServer()
 		require.NoError(t, err)
 	})
-}
-
-func initFakeServer() (*httptest.Server, error) {
-	smHandler, err := newFakeSMHandler()
-	if err != nil {
-		return nil, fmt.Errorf("while creating new fake SM handler: %w", err)
-	}
-
-	mux := http.NewServeMux()
-	mux.HandleFunc("GET /v1/service_offerings", smHandler.getServiceOfferings)
-	mux.HandleFunc("GET /v1/service_offerings/{serviceOfferingID}", smHandler.getServiceOffering)
-	mux.HandleFunc("GET /v1/service_plans", smHandler.getServicePlans)
-	mux.HandleFunc("GET /v1/service_instances", smHandler.getServiceInstances)
-	mux.HandleFunc("GET /v1/service_instances/{serviceInstanceID}", smHandler.getServiceInstance)
-	mux.HandleFunc("POST /v1/service_instances", smHandler.createServiceInstance)
-	mux.HandleFunc("PATCH /v1/service_instances/{serviceInstanceID}", smHandler.updateServiceInstance)
-	mux.HandleFunc("DELETE /v1/service_instances/{serviceInstanceID}", smHandler.deleteServiceInstance)
-	mux.HandleFunc("GET /v1/service_bindings", smHandler.getServiceBindings)
-	mux.HandleFunc("GET /v1/service_bindings/{serviceBindingID}", smHandler.getServiceBinding)
-	mux.HandleFunc("POST /v1/service_bindings", smHandler.createServiceBinding)
-	mux.HandleFunc("DELETE /v1/service_bindings/{serviceBindingID}", smHandler.deleteServiceBinding)
-
-	srv := httptest.NewUnstartedServer(mux)
-
-	return srv, nil
-}
-
-type fakeSMHandler struct {
-	serviceOfferings *types.ServiceOfferings
-	servicePlans     *types.ServicePlans
-	serviceInstances *types.ServiceInstances
-	serviceBindings  *types.ServiceBindings
-}
-
-func newFakeSMHandler() (*fakeSMHandler, error) {
-	sos, err := getServiceOfferingsFromJSON()
-	if err != nil {
-		return nil, fmt.Errorf("while getting service offerings from JSON: %w", err)
-
-	}
-	plans, err := getServicePlansFromJSON()
-	if err != nil {
-		return nil, fmt.Errorf("while getting service plans from JSON: %w", err)
-	}
-	sis, err := getServiceInstancesFromJSON()
-	if err != nil {
-		return nil, fmt.Errorf("while getting service instances from JSON: %w", err)
-
-	}
-	sbs, err := getServiceBindingsFromJSON()
-	if err != nil {
-		return nil, fmt.Errorf("while getting service bindings from JSON: %w", err)
-
-	}
-	return &fakeSMHandler{serviceOfferings: sos, servicePlans: plans, serviceInstances: sis, serviceBindings: sbs}, nil
-}
-
-func getServiceOfferingsFromJSON() (*types.ServiceOfferings, error) {
-	var sos types.ServiceOfferings
-	f, err := os.Open(serviceOfferingsJSONPath)
-	defer f.Close()
-	if err != nil {
-		return nil, fmt.Errorf("while reading resources from JSON file: %w", err)
-	}
-
-	d := json.NewDecoder(f)
-	if err := d.Decode(&sos); err != nil {
-		return nil, fmt.Errorf("while decoding resources JSON: %w", err)
-	}
-	return &sos, nil
-}
-
-func getServicePlansFromJSON() (*types.ServicePlans, error) {
-	var sps types.ServicePlans
-	f, err := os.Open(servicePlansJSONPath)
-	defer f.Close()
-	if err != nil {
-		return nil, fmt.Errorf("while reading resources from JSON file: %w", err)
-	}
-
-	d := json.NewDecoder(f)
-	if err := d.Decode(&sps); err != nil {
-		return nil, fmt.Errorf("while decoding resources JSON: %w", err)
-	}
-
-	return &sps, nil
-}
-
-func getServiceInstancesFromJSON() (*types.ServiceInstances, error) {
-	var sis types.ServiceInstances
-	f, err := os.Open(serviceInstancesJSONPath)
-	defer f.Close()
-	if err != nil {
-		return nil, fmt.Errorf("while reading resources from JSON file: %w", err)
-	}
-
-	d := json.NewDecoder(f)
-	if err := d.Decode(&sis); err != nil {
-		return nil, fmt.Errorf("while decoding resources JSON: %w", err)
-	}
-
-	return &sis, nil
-}
-
-func getServiceBindingsFromJSON() (*types.ServiceBindings, error) {
-	var sbs types.ServiceBindings
-	f, err := os.Open(serviceBindingsJSONPath)
-	defer f.Close()
-	if err != nil {
-		return nil, fmt.Errorf("while reading resources from JSON file: %w", err)
-	}
-
-	d := json.NewDecoder(f)
-	if err := d.Decode(&sbs); err != nil {
-		return nil, fmt.Errorf("while decoding resources JSON: %w", err)
-	}
-
-	return &sbs, nil
-}
-
-func (h *fakeSMHandler) getServiceOfferings(w http.ResponseWriter, r *http.Request) {
-	data, err := json.Marshal(h.serviceOfferings)
-	if err != nil {
-		log.Println("error while marshalling service offerings data: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service offerings data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) getServiceOffering(w http.ResponseWriter, r *http.Request) {
-	soID := r.PathValue("serviceOfferingID")
-	if len(soID) == 0 {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	var err error
-	data := make([]byte, 0)
-	for _, so := range h.serviceOfferings.Items {
-		if so.ID == soID {
-			data, err = json.Marshal(so)
-			if err != nil {
-				log.Println("error while marshalling service offering data: %w", err)
-				w.WriteHeader(http.StatusInternalServerError)
-				return
-			}
-			w.WriteHeader(http.StatusOK)
-			break
-		}
-		w.WriteHeader(http.StatusNotFound)
-	}
-
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service offerings data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) getServicePlans(w http.ResponseWriter, r *http.Request) {
-	values := r.URL.Query()
-	prefixedSoID := values.Get(servicemanager.URLFieldQueryKey)
-	IDFilter := ""
-	if len(prefixedSoID) != 0 {
-		fields := strings.Fields(prefixedSoID)
-		IDFilter = strings.Trim(fields[2], "'")
-	}
-
-	var responseSps types.ServicePlans
-	if len(IDFilter) != 0 {
-		var filteredSps types.ServicePlans
-		for _, sp := range h.servicePlans.Items {
-			if sp.ServiceOfferingID == IDFilter {
-				filteredSps.Items = append(filteredSps.Items, sp)
-			}
-		}
-		responseSps = filteredSps
-	}
-
-	data, err := json.Marshal(responseSps)
-	if err != nil {
-		log.Println("error while marshalling service plans data: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service plans data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) getServiceInstances(w http.ResponseWriter, r *http.Request) {
-	data, err := json.Marshal(h.serviceInstances)
-	if err != nil {
-		log.Println("error while marshalling service instances data: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service instances data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) getServiceInstance(w http.ResponseWriter, r *http.Request) {
-	siID := r.PathValue("serviceInstanceID")
-	if len(siID) == 0 {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	var err error
-	data := make([]byte, 0)
-	for _, si := range h.serviceInstances.Items {
-		if si.ID == siID {
-			data, err = json.Marshal(si)
-			if err != nil {
-				log.Println("error while marshalling service instance data: %w", err)
-				w.WriteHeader(http.StatusInternalServerError)
-				return
-			}
-			w.WriteHeader(http.StatusOK)
-			break
-		}
-	}
-	if len(data) == 0 {
-		errResp := types.ErrorResponse{
-			ErrorType:   "NotFound",
-			Description: "could not find such service_instance",
-		}
-		data, err = json.Marshal(errResp)
-		if err != nil {
-			log.Println("error while marshalling error response: %w", err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusNotFound)
-	}
-
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service instance data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) createServiceInstance(w http.ResponseWriter, r *http.Request) {
-	var siCreateRequest types.ServiceInstance
-	err := json.NewDecoder(r.Body).Decode(&siCreateRequest)
-	if err != nil {
-		log.Println("error while decoding request body into Service Instance struct: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	siCreateRequest.ID = uuid.New().String()
-	h.serviceInstances.Items = append(h.serviceInstances.Items, siCreateRequest)
-
-	data, err := json.Marshal(siCreateRequest)
-	if err != nil {
-		log.Println("error while marshalling service instance: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusCreated)
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service instance data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) updateServiceInstance(w http.ResponseWriter, r *http.Request) {
-	siID := r.PathValue("serviceInstanceID")
-	if len(siID) == 0 {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	var siUpdateRequest types.ServiceInstanceUpdateRequest
-	err := json.NewDecoder(r.Body).Decode(&siUpdateRequest)
-	if err != nil {
-		log.Println("error while decoding request body into ServiceInstanceUpdateRequest struct: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	var siUpdateResponse *types.ServiceInstance
-	for i, si := range h.serviceInstances.Items {
-		if si.ID == siID {
-			siUpdateResponse = &h.serviceInstances.Items[i]
-			if siUpdateRequest.Name != nil {
-				h.serviceInstances.Items[i].Name = *siUpdateRequest.Name
-			}
-			if siUpdateRequest.ServicePlanID != nil {
-				h.serviceInstances.Items[i].ServicePlanID = *siUpdateRequest.ServicePlanID
-			}
-			if siUpdateRequest.Shared != nil {
-				h.serviceInstances.Items[i].Shared = *siUpdateRequest.Shared
-			}
-			if siUpdateRequest.Parameters != nil {
-				h.serviceInstances.Items[i].Parameters = *siUpdateRequest.Parameters
-			}
-			if len(siUpdateRequest.Labels) != 0 {
-				for _, labelChange := range siUpdateRequest.Labels {
-					if labelChange.Operation == types.AddLabelOperation {
-						h.serviceInstances.Items[i].Labels[labelChange.Key] = labelChange.Values
-					} else if labelChange.Operation == types.RemoveLabelOperation {
-						delete(h.serviceInstances.Items[i].Labels, labelChange.Key)
-					}
-				}
-			}
-			break
-		}
-	}
-
-	data, err := json.Marshal(siUpdateResponse)
-	if err != nil {
-		log.Println("error while marshalling service instance: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service instance data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) deleteServiceInstance(w http.ResponseWriter, r *http.Request) {
-	siID := r.PathValue("serviceInstanceID")
-	if len(siID) == 0 {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	for i, si := range h.serviceInstances.Items {
-		if si.ID == siID {
-			h.serviceInstances.Items = append(h.serviceInstances.Items[:i], h.serviceInstances.Items[i+1:]...)
-			w.WriteHeader(http.StatusOK)
-			return
-		}
-	}
-	w.WriteHeader(http.StatusNotFound)
-
-	errResp := types.ErrorResponse{
-		ErrorType:   "NotFound",
-		Description: "could not find such service_instance",
-	}
-
-	data, err := json.Marshal(errResp)
-	if err != nil {
-		log.Println("error while marshalling error response: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing error response data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) getServiceBindings(w http.ResponseWriter, r *http.Request) {
-	data, err := json.Marshal(h.serviceBindings)
-	if err != nil {
-		log.Println("error while marshalling service bindings data: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusOK)
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service bindings data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) getServiceBinding(w http.ResponseWriter, r *http.Request) {
-	sbID := r.PathValue("serviceBindingID")
-	if len(sbID) == 0 {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	var err error
-	data := make([]byte, 0)
-	for _, sb := range h.serviceBindings.Items {
-		if sb.ID == sbID {
-			data, err = json.Marshal(sb)
-			if err != nil {
-				log.Println("error while marshalling service binding data: %w", err)
-				w.WriteHeader(http.StatusInternalServerError)
-				return
-			}
-			w.WriteHeader(http.StatusOK)
-			break
-		}
-	}
-	if len(data) == 0 {
-		errResp := types.ErrorResponse{
-			ErrorType:   "NotFound",
-			Description: "could not find such service_binding",
-		}
-		data, err = json.Marshal(errResp)
-		if err != nil {
-			log.Println("error while marshalling error response: %w", err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusNotFound)
-	}
-
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service binding data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) createServiceBinding(w http.ResponseWriter, r *http.Request) {
-	var sbCreateRequest types.ServiceBinding
-	err := json.NewDecoder(r.Body).Decode(&sbCreateRequest)
-	if err != nil {
-		log.Println("error while decoding request body into Service Binding struct: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	sbCreateRequest.ID = uuid.New().String()
-	h.serviceBindings.Items = append(h.serviceBindings.Items, sbCreateRequest)
-
-	data, err := json.Marshal(sbCreateRequest)
-	if err != nil {
-		log.Println("error while marshalling service binding: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	w.WriteHeader(http.StatusCreated)
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing service binding data: %w", err)
-		return
-	}
-}
-
-func (h *fakeSMHandler) deleteServiceBinding(w http.ResponseWriter, r *http.Request) {
-	sbID := r.PathValue("serviceBindingID")
-	if len(sbID) == 0 {
-		w.WriteHeader(http.StatusBadRequest)
-		return
-	}
-
-	for i, sb := range h.serviceBindings.Items {
-		if sb.ID == sbID {
-			h.serviceBindings.Items = append(h.serviceBindings.Items[:i], h.serviceBindings.Items[i+1:]...)
-			w.WriteHeader(http.StatusOK)
-			return
-		}
-	}
-	w.WriteHeader(http.StatusNotFound)
-
-	errResp := types.ErrorResponse{
-		ErrorType:   "NotFound",
-		Description: "could not find such service_binding",
-	}
-
-	data, err := json.Marshal(errResp)
-	if err != nil {
-		log.Println("error while marshalling error response: %w", err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	if _, err = w.Write(data); err != nil {
-		w.WriteHeader(http.StatusInternalServerError)
-		log.Println("error while writing error response data: %w", err)
-		return
-	}
-}
-
-type fakeSecretProvider struct {
-	secrets []*corev1.Secret
-}
-
-func newFakeSecretProvider() *fakeSecretProvider {
-	return &fakeSecretProvider{secrets: make([]*corev1.Secret, 0)}
-}
-
-func (p *fakeSecretProvider) AddSecret(secret *corev1.Secret) {
-	p.secrets = append(p.secrets, secret)
-}
-
-func (p *fakeSecretProvider) GetByNameAndNamespace(ctx context.Context, name, namespace string) (*corev1.Secret, error) {
-	for _, secret := range p.secrets {
-		if secret.Name == name && secret.Namespace == namespace {
-			return secret, nil
-		}
-	}
-	return nil, fmt.Errorf("secret not found")
-}
-
-func (p *fakeSecretProvider) clean() {
-	p.secrets = make([]*corev1.Secret, 0)
-}
-
-func defaultSecret() *corev1.Secret {
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "sap-btp-service-operator",
-			Namespace: "kyma-system",
-		},
-		StringData: map[string]string{
-			"clientid":       "default-client-id",
-			"clientsecret":   "default-client-secret",
-			"sm_url":         "https://default-sm-url.local",
-			"tokenurl":       "https://default-token-url.local",
-			"tokenurlsuffix": "/oauth/token",
-		},
-	}
 }
 
 func getServiceOfferingByID(serviceOfferings *types.ServiceOfferings, serviceOfferingID string) *types.ServiceOffering {
